requirements.txt
streamlit
openai
python-dotenv
httpx
PyMuPDF
python-docx
python-pptx
pandas
Pillow
tavily-python